[Kernels][GPU] Support TMA Multicast (#54564)
H100 TMA support multicasting a loading a tile from global memory to the
shared memory's in a Thread Block clusters.

GPU_ORIG_REV_ID: 5c76669062f38cf46b66ec34e815f9cb9346b59e

+++ mojo/stdlib/src/gpu/id.mojo
@@ -50,8 +50,14 @@ fn sm_id() -> UInt:
 #  1D ctaid in a cluster
 # ===-----------------------------------------------------------------------===#
 @always_inline
-fn block_rank_in_cluster() -> UInt32:
+fn block_rank_in_cluster() -> UInt:
     """Get 1D ctaid in a cluster."""
-    return llvm_intrinsic[
-        "llvm.nvvm.read.ptx.sreg.cluster.ctarank", UInt32, has_side_effect=False
-    ]()
+    return UInt(
+        Int(
+            llvm_intrinsic[
+                "llvm.nvvm.read.ptx.sreg.cluster.ctarank",
+                UInt32,
+                has_side_effect=False,
+            ]()
+        )
+    )
